refactor(website): moved hooks and assets
