+++ aksel.nav.no/website/app/(routes)/(designsystemet)/designsystemet/_ui/change-log/ChangeLogNews.tsx
@@ -15,9 +15,9 @@ import {
 import { sanityFetch } from "@/app/_sanity/live";
 import { N_LATEST_CHANGE_LOGS_QUERY } from "@/app/_sanity/queries";
 import type { N_LATEST_CHANGE_LOGS_QUERY_RESULT } from "@/app/_sanity/query-types";
+import { FigmaIcon, GithubIcon } from "@/app/_ui/assets/Icons";
 import { MarkdownText } from "@/app/_ui/typography/MarkdownText";
 import { UmamiLink } from "@/app/_ui/umami/UmamiLink";
-import { FigmaIcon, GithubIcon } from "@/assets/Icons";
 
 type ChangeLogNewsProps = {
   title: string;
