+++ aksel.nav.no/website/app/_ui/toc/TableOfContents.tsx
@@ -4,9 +4,9 @@ import { stegaClean } from "next-sanity";
 import { Events } from "@navikt/analytics-types";
 import { BodyShort, Button, Detail } from "@navikt/ds-react";
 import type { TOC_BY_SLUG_QUERY_RESULT } from "@/app/_sanity/query-types";
+import { GithubIcon } from "@/app/_ui/assets/Icons";
 import { NextLink } from "@/app/_ui/next-link/NextLink";
 import { umamiTrack } from "@/app/_ui/umami/Umami.track";
-import { GithubIcon } from "@/assets/Icons";
 import { cl } from "@/ui-utils/className";
 import { removeEmojiesFromText } from "@/ui-utils/format-text";
 import styles from "./TableOfContents.module.css";
