+++ aksel.nav.no/website/app/_ui/utils/hooks/useEscapeKeydown.ts
